@@ -1,100 +1,111 @@
 {
-    "name": "jy-transform",
-    "description": "This project aims to read, write and transform YAML, JS or JSON objects into each other using CLI or API, while the source and destination resources can be files on CLI and additionally, objects or streams on API level.",
-    "version": "2.0.1",
-    "homepage": "https://github.com/deadratfink/jy-transform",
-    "author": {
-        "name": "Jens Krefeldt",
-        "email": "j.krefeldt@gmail.com",
-        "url": "https://github.com/deadratfink"
-    },
-    "contributors": [
-    ],
-    "license": "SEE LICENSE IN [LICENSE.md](https://github.com/deadratfink/jy-transform/blob/master/LICENSE.md)",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/deadratfink/jy-transform.git"
-    },
-    "publishConfig": {
-        "registry": "https://registry.npmjs.org/"
-    },
-    "bugs": "https://github.com/deadratfink/jy-transform/issues",
-    "private": false,
-    "config": {
-        "test": {
-            "mocha": {
-                "unit": {
-                    "reporter": "spec"
-                }
-            }
-        }
-    },
-    "scripts": {
-        "docs": "cat docs/LOGO.md > README.md && cat docs/BADGES.md >> README.md && cat docs/TOC.md >> README.md && package-json-to-readme --no-footer ./package.json >> README.md && cat docs/USAGE.md >> README.md && echo '\n# Changelog\n' >> README.md && cat docs/CHANGELOG.md >> README.md && doctoc README.md --github --title '# TOC' --maxlevel 2",
-        "wiki": "jsdoc2md ./jyt lib/*.js index.js > docs/API.md && doctoc docs/API.md --github --title '### TOC' --maxlevel 2 && cat docs/API.md > '../jy-transform.wiki/API-v2.md' && cat docs/CONTRIBUTING.md > ../jy-transform.wiki/Contributing.md && cat docs/CHANGELOG.md > ../jy-transform.wiki/Changelog.md && doctoc ../jy-transform.wiki/Changelog.md --github --title '### TOC' --maxlevel 3",
-        "pretest": "rm -Rf test/tmp",
-        "test": "istanbul cover _mocha --report lcovonly -- -R $npm_package_config_test_mocha_unit_reporter ./test/test*.js"
-    },
-    "engines": {
-        "node": ">=0.10.0"
-    },
-    "dependencies": {
-        "bluebird": "^3.4.6",
-        "cli": "^1.0.1",
-        "is-stream": "^1.1.0",
-        "js-yaml": "^3.6.1",
-        "json-stringify-safe": "^5.0.1",
-        "mkdirp-then": "^1.2.0",
-        "serialize-js": "^1.1.0"
-    },
-    "devDependencies": {
-        "codeclimate-test-reporter": "^0.4.0",
-        "codecov": "^1.0.1",
-        "coveralls": "^2.11.14",
-        "doctoc": "^1.0.0",
-        "fs-extra": "^1.0.0",
-        "istanbul": "^0.4.5",
-        "jsdoc-parse": "^2.0.5",
-        "jsdoc-to-markdown": "^2.0.1",
-        "mocha": "^3.1.2",
-        "mocha-lcov-reporter": "^1.2.0",
-        "object-path": "^0.11.2",
-        "package-json-to-readme": "^1.5.1",
-        "winston": "^2.3.0"
-    },
-    "preferGlobal": true,
-    "bin": {
-        "jyt": "./jyt"
-    },
-    "main": "./index.js",
-    "keywords": [
-        "api",
-        "cli",
-        "jy",
-        "jyt",
-        "jy-transform",
-        "transform",
-        "convert",
-        "javascript",
-        "js",
-        "json",
-        "yaml",
-        "yaml2js",
-        "yaml-2-js",
-        "yaml2json",
-        "yaml-2-json",
-        "js2yaml",
-        "js-2-yaml",
-        "json2yaml",
-        "json-2-yaml",
-        "yamltojs",
-        "yaml-to-js",
-        "yamltojson",
-        "yaml-to-json",
-        "jstoyaml",
-        "js-to-yaml",
-        "jsontoyaml",
-        "json-to-yaml",
-        "promise"
-    ]
+  "name": "jy-transform",
+  "description": "This project aims to read, write and transform YAML, JS or JSON objects into each other using CLI or API, while the source and destination resources can be files on CLI and additionally, objects or streams on API level.",
+  "version": "3.0.0",
+  "homepage": "https://github.com/deadratfink/jy-transform",
+  "author": {
+    "name": "Jens Krefeldt",
+    "email": "j.krefeldt@gmail.com",
+    "url": "https://github.com/deadratfink"
+  },
+  "contributors": [],
+  "license": "SEE LICENSE IN [LICENSE.md](https://github.com/deadratfink/jy-transform/blob/master/LICENSE.md)",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/deadratfink/jy-transform.git"
+  },
+  "publishConfig": {
+    "registry": "https://registry.npmjs.org/"
+  },
+  "bugs": "https://github.com/deadratfink/jy-transform/issues",
+  "private": false,
+  "scripts": {
+    "build": "babel src -d lib",
+    "docs": "cat docs/LOGO.md > README.md && cat docs/BADGES.md >> README.md && cat docs/TOC.md >> README.md && package-json-to-readme --no-footer ./package.json >> README.md && cat docs/USAGE.md >> README.md && echo '\n# Changelog\n' >> README.md && cat docs/CHANGELOG.md >> README.md && doctoc README.md --github --title '# TOC' --maxlevel 2",
+    "wiki": "jsdoc2md ./jyt lib/*.js index.js > docs/API.md && doctoc docs/API.md --github --title '### TOC' --maxlevel 2 && cat docs/API.md > '../jy-transform.wiki/API-v2.md' && cat docs/CONTRIBUTING.md > ../jy-transform.wiki/Contributing.md && cat docs/CHANGELOG.md > ../jy-transform.wiki/Changelog.md && doctoc ../jy-transform.wiki/Changelog.md --github --title '### TOC' --maxlevel 3",
+    "pretestBAK": "rm -Rf test/tmp",
+    "test": "jest --no-cache --coverage --runInBand --config=./.jestrc.js",
+    "eslint": "eslint ."
+  },
+  "engines": {
+    "node": ">=4.0.0"
+  },
+  "dependencies": {
+    "bluebird": "^3.4.6",
+    "cli": "^1.0.1",
+    "is-stream": "^1.1.0",
+    "js-yaml": "^3.6.1",
+    "json-stringify-safe": "^5.0.1",
+    "mkdirp-then": "^1.2.0",
+    "promisify-es6": "~1.0.2",
+    "serialize-js": "^1.1.0"
+  },
+  "devDependencies": {
+    "babel-cli": "~6.24.1",
+    "babel-core": "~6.24.1",
+    "babel-eslint": "~7.2.3",
+    "babel-plugin-transform-builtin-extend": "~1.1.2",
+    "babel-preset-env": "~1.4.0",
+    "babel-preset-stage-0": "~6.24.1",
+    "babel-watch": "~2.0.6",
+    "chalk": "~1.1.3",
+    "codeclimate-test-reporter": "^0.4.0",
+    "codecov": "^1.0.1",
+    "combarnea-winston-console-formatter": "~1.0.1",
+    "coveralls": "^2.11.14",
+    "doctoc": "^1.0.0",
+    "eslint": "~3.19.0",
+    "eslint-config-airbnb-base": "~10.0.1",
+    "eslint-plugin-filenames": "~1.2.0",
+    "eslint-plugin-import": "~2.2.0",
+    "eslint-plugin-jest": "~20.0.3",
+    "eslint-plugin-jest-async": "~1.0.3",
+    "eslint-plugin-jsdoc": "~2.3.1",
+    "fs-extra": "^1.0.0",
+    "istanbul": "^0.4.5",
+    "jest": "~20.0.4",
+    "jsdoc-parse": "^2.0.5",
+    "jsdoc-to-markdown": "^2.0.1",
+    "mocha": "^3.1.2",
+    "mocha-lcov-reporter": "^1.2.0",
+    "object-path": "^0.11.2",
+    "package-json-to-readme": "^1.5.1",
+    "winston": "^2.3.0",
+    "winston-console-formatter": "~0.3.1"
+  },
+  "preferGlobal": true,
+  "bin": {
+    "jyt": "./jyt"
+  },
+  "main": "./index.js",
+  "keywords": [
+    "api",
+    "cli",
+    "jy",
+    "jyt",
+    "jy-transform",
+    "transform",
+    "convert",
+    "javascript",
+    "js",
+    "json",
+    "yaml",
+    "yaml2js",
+    "yaml-2-js",
+    "yaml2json",
+    "yaml-2-json",
+    "js2yaml",
+    "js-2-yaml",
+    "json2yaml",
+    "json-2-yaml",
+    "yamltojs",
+    "yaml-to-js",
+    "yamltojson",
+    "yaml-to-json",
+    "jstoyaml",
+    "js-to-yaml",
+    "jsontoyaml",
+    "json-to-yaml",
+    "promise"
+  ]
 }
