@@ -1,7 +1,8 @@
-'use strict';
+module.exports = {
+  Transformer: require('./lib/transformer.js'),
+  Reader: require('./lib/reader.js'),
+  Writer: require('./lib/writer.js'),
+  Constants: require('./lib/constants.js'),
+  Middleware: require('./lib/middleware.js'),
+};
 
-module.exports = module.exports.Transformer = require('./lib/transformer.js');
-module.exports.Reader = require('./lib/reader.js');
-module.exports.Writer = require('./lib/writer.js');
-module.exports.middleware = require('./lib/middleware.js');
-module.exports.constants = require('./lib/constants.js');
