@@ -1,8 +1,8 @@
 'use strict';
 
-///////////////////////////////////////////////////////////////////////////////
+// ////////////////////////////////////////////////////////////////////////////
 // CONSTRUCTOR
-///////////////////////////////////////////////////////////////////////////////
+// ////////////////////////////////////////////////////////////////////////////
 
 /**
  * Constructs the constants.
@@ -12,7 +12,7 @@
  * @class Class which defines all constants usable in or with this module.
  */
 function Constants() {
-    return this;
+  return this;
 }
 
 Constants.prototype = {};
@@ -20,9 +20,9 @@ Constants.prototype.constructor = Constants;
 var constants = new Constants();
 module.exports = constants;
 
-///////////////////////////////////////////////////////////////////////////////
+// ////////////////////////////////////////////////////////////////////////////
 // CONSTANTS
-///////////////////////////////////////////////////////////////////////////////
+// ////////////////////////////////////////////////////////////////////////////
 
 /**
  * The 'utf8' constant.
@@ -187,13 +187,13 @@ Constants.prototype.DEFAULT_JS_EXPORTS_IDENTIFIER = undefined;
  * @see {@link DEST_DESCRIPTION}
  */
 Constants.prototype.DEFAULT_OPTIONS = {
-    origin:  constants.ORIGIN_DESCRIPTION,
-    target:  constants.TARGET_DESCRIPTION,
-    dest:    constants.DEST_DESCRIPTION,
-    indent:  constants.DEFAULT_INDENT,
-    force:   constants.DEFAULT_FORCE_FILE_OVERWRITE,
-    imports: constants.DEFAULT_JS_IMPORTS_IDENTIFIER,
-    exports: constants.DEFAULT_JS_EXPORTS_IDENTIFIER
+  origin:  constants.ORIGIN_DESCRIPTION,
+  target:  constants.TARGET_DESCRIPTION,
+  dest:    constants.DEST_DESCRIPTION,
+  indent:  constants.DEFAULT_INDENT,
+  force:   constants.DEFAULT_FORCE_FILE_OVERWRITE,
+  imports: constants.DEFAULT_JS_IMPORTS_IDENTIFIER,
+  exports: constants.DEFAULT_JS_EXPORTS_IDENTIFIER
 };
 
 /**
@@ -285,13 +285,13 @@ Constants.prototype.JS_TO_JS = 'js2js';
  * @public
  */
 Constants.prototype.TRANSFORMATIONS = [
-    constants.YAML_TO_JS,
-    constants.YAML_TO_JSON,
-    constants.JS_TO_YAML,
-    constants.JSON_TO_YAML,
-    constants.JSON_TO_JS,
-    constants.JS_TO_JSON,
-    constants.YAML_TO_YAML,
-    constants.JSON_TO_JSON,
-    constants.JS_TO_JS
+  constants.YAML_TO_JS,
+  constants.YAML_TO_JSON,
+  constants.JS_TO_YAML,
+  constants.JSON_TO_YAML,
+  constants.JSON_TO_JS,
+  constants.JS_TO_JSON,
+  constants.YAML_TO_YAML,
+  constants.JSON_TO_JSON,
+  constants.JS_TO_JS
 ];
