@@ -1,10 +1,8 @@
-'use strict';
-
 var Promise = require('bluebird');
 
-///////////////////////////////////////////////////////////////////////////////
+// ////////////////////////////////////////////////////////////////////////////
 // CONSTRUCTOR
-///////////////////////////////////////////////////////////////////////////////
+// ////////////////////////////////////////////////////////////////////////////
 
 /**
  * Constructs the `LogWrapper`.
@@ -22,21 +20,21 @@ var Promise = require('bluebird');
  */
 function LogWrapper(logger) {
 
-    /**
-     * The logger instance.
-     *
-     * @member {(logger|cli|console)}
-     * @private
-     */
-    this.logInstance = logger || console;
+  /**
+   * The logger instance.
+   *
+   * @member {(logger|cli|console)}
+   * @private
+   */
+  this.logInstance = logger || console;
 }
 
 LogWrapper.prototype = {};
 LogWrapper.prototype.constructor = LogWrapper;
 
-///////////////////////////////////////////////////////////////////////////////
+// ////////////////////////////////////////////////////////////////////////////
 // LOGGER METHODS
-///////////////////////////////////////////////////////////////////////////////
+// ////////////////////////////////////////////////////////////////////////////
 
 /**
  * Log the options with INFO level.
@@ -50,7 +48,7 @@ LogWrapper.prototype.constructor = LogWrapper;
  * @public
  */
 LogWrapper.prototype.info = function (msg) {
-    this.logInstance.info(msg);
+  this.logInstance.info(msg);
 };
 
 /**
@@ -65,11 +63,11 @@ LogWrapper.prototype.info = function (msg) {
  * @public
  */
 LogWrapper.prototype.debug = function (msg) {
-    if (this.logInstance.debug && typeof this.logInstance.debug === 'function') {
-        this.logInstance.debug(msg);
-    } else {
-        this.info(msg);
-    }
+  if (this.logInstance.debug && typeof this.logInstance.debug === 'function') {
+    this.logInstance.debug(msg);
+  } else {
+    this.info(msg);
+  }
 };
 
 /**
@@ -85,11 +83,11 @@ LogWrapper.prototype.debug = function (msg) {
  * @see {@link #debug}
  */
 LogWrapper.prototype.trace = function (msg) {
-    if (this.logInstance.trace && typeof this.logInstance.trace === 'function') {
-        this.logInstance.trace(msg);
-    } else {
-        this.debug(msg);
-    }
+  if (this.logInstance.trace && typeof this.logInstance.trace === 'function') {
+    this.logInstance.trace(msg);
+  } else {
+    this.debug(msg);
+  }
 };
 
 /**
@@ -104,7 +102,7 @@ LogWrapper.prototype.trace = function (msg) {
  * @public
  */
 LogWrapper.prototype.error = function (msg) {
-    this.logInstance.error(msg);
+  this.logInstance.error(msg);
 };
 
 /**
@@ -125,16 +123,16 @@ LogWrapper.prototype.error = function (msg) {
  * @public
  */
 LogWrapper.prototype.verboseOptions = function (options) {
-    var self = this;
-    return Promise.resolve()
-        .then(function () {
-            self.info('origin: ' + options.origin);
-            self.info('target: ' + options.target);
-            self.info('src:    ' + options.src);
-            self.info('dest:   ' + options.dest);
-            self.info('indent: ' + options.indent);
-            return options;
-        });
+  var self = this;
+  return Promise.resolve()
+    .then(function () {
+      self.info('origin: ' + options.origin);
+      self.info('target: ' + options.target);
+      self.info('src:    ' + options.src);
+      self.info('dest:   ' + options.dest);
+      self.info('indent: ' + options.indent);
+      return options;
+    });
 };
 
 exports = module.exports = LogWrapper;
