@@ -1,10 +1,8 @@
-'use strict';
-
 var Promise = require('bluebird');
 
-///////////////////////////////////////////////////////////////////////////////
+// ////////////////////////////////////////////////////////////////////////////
 // CONSTRUCTOR
-///////////////////////////////////////////////////////////////////////////////
+// ////////////////////////////////////////////////////////////////////////////
 
 /**
  * Constructs the `Middleware`.
@@ -29,7 +27,7 @@ module.exports = new Middleware();
  * @private
  */
 function identity(object) {
-    return Promise.resolve(object);
+  return Promise.resolve(object);
 }
 
 /**
@@ -61,9 +59,8 @@ Middleware.prototype.identityMiddleware = identity;
  *        ```
  *
  *        **NOTE:** the Promise has to return the processed JSON!
- * @returns {Promise}  - The given middleware Promise or a new JSON 'identity' middleware Promise.
- * @throws {TypeError} - Will throw this error when the passed `middleware`
- *         is not type of `Function`.
+ * @returns {Function} - The given middleware Promise or a new JSON 'identity' middleware Promise function.
+ * @throws {TypeError} - Will throw this error when the passed `middleware` is not type of `Function`.
  * @example
  * var middleware = require('./lib/middleware.js');
  * var myMiddleware = function(object) {
@@ -76,11 +73,11 @@ Middleware.prototype.identityMiddleware = identity;
  * @public
  */
 Middleware.prototype.ensureMiddleware = function (middleware) {
-    if (middleware !== undefined && (typeof middleware !== 'function')) {
-        return Promise.reject(new TypeError('The provided middleware is not a Function type'));
-    }
-    if (!middleware) {
-        middleware = identity;
-    }
-    return middleware;
+  if (middleware !== undefined && (typeof middleware !== 'function')) {
+    throw new TypeError('The provided middleware is not a Function type');
+  }
+  if (!middleware) {
+    middleware = identity;
+  }
+  return middleware;
 };
